+++ src/Phan/CLI.php
@@ -38,7 +38,7 @@ class CLI
      * still available: g,n,t,u,w
      * @internal
      */
-    const GETOPT_SHORT_OPTIONS = 'f:m:o:c:k:aeqbr:pid:3:y:l:xj:zhvs:';
+    const GETOPT_SHORT_OPTIONS = 'f:m:o:c:k:aeqbr:pid:3:y:l:xj:zhvs:SCP:';
 
     /**
      * List of long flags passed to getopt
@@ -411,6 +411,7 @@ class CLI
                     // Slightly faster, e.g. for daemon mode with lowest latency (along with --quick).
                     Config::setValue('plugins', []);
                     break;
+                case 'P':
                 case 'plugin':
                     if (!is_array($value)) {
                         $value = [$value];
@@ -435,6 +436,7 @@ class CLI
                 case 'strict-return-checking':
                     Config::setValue('strict_return_checking', true);
                     break;
+                case 'S':
                 case 'strict-type-checking':
                     Config::setValue('strict_method_checking', true);
                     Config::setValue('strict_param_checking', true);
@@ -548,6 +550,7 @@ class CLI
                 case 'markdown-issue-messages':
                     Config::setValue('markdown_issue_messages', true);
                     break;
+                case 'C':
                 case 'color':
                     Config::setValue('color_issue_messages', true);
                     break;
@@ -787,7 +790,7 @@ Usage: {$argv[0]} [options] [files...]
   Output filename
 
 $init_help
- --color
+ -C, --color
   Add colors to the outputted issues. Tested in Unix.
   This is recommended for only the default --output-mode ('text')
 
@@ -838,23 +841,31 @@ $init_help
  --disable-plugins
   Don't run any plugins. Slightly faster.
 
- --plugin <pluginName|path/to/Plugin.php>
+ -P, --plugin <pluginName|path/to/Plugin.php>
   Add a plugin to run. This flag can be repeated.
   (Either pass the name of the plugin or a relative/absolute path to the plugin)
 
  --strict-method-checking
-  Enables the config option `strict_method_checking`.
+  Warn if any type in a method invocation's object is definitely not an object,
+  or any type in an invoked expression is not a callable.
+  (Enables the config option `strict_method_checking`)
 
  --strict-param-checking
-  Enables the config option `strict_param_checking`.
+  Warn if any type in an argument's union type cannot be cast to
+  the parameter's expected union type.
+  (Enables the config option `strict_param_checking`)
 
  --strict-property-checking
-  Enables the config option `strict_property_checking`.
+  Warn if any type in a property assignment's union type
+  cannot be cast to a type in the property's declared union type.
+  (Enables the config option `strict_property_checking`)
 
  --strict-return-checking
-  Enables the config option `strict_return_checking`.
+  Warn if any type in a returned value's union type
+  cannot be cast to the declared return type.
+  (Enables the config option `strict_return_checking`)
 
- --strict-type-checking
+ -S, --strict-type-checking
   Equivalent to
   `--strict-method-checking --strict-param-checking --strict-property-checking --strict-return-checking`.