+++ NEWS.md
@@ -3,6 +3,9 @@ Phan NEWS
 ?? ??? 2019, Phan 1.2.1 (dev)
 -----------------------
 
+New features(CLI):
++ Add short flags: `-S` for `--strict-type-checking`, `-C` for `--color`, `-P` for `--plugin <plugin>`
+
 New features(Analysis):
 + Infer that the result of `array_map` has integer keys when passed two or more arrays (#2277)
 + Improve inferences about the left hand side of `&&` statements such as `$leftVar && (other_expression);` (#2300)